Create test.py
This is not a test that can be incorporated in an  automatic test suite. It can be used manually, however, to verify the behaviour of the module. While this is of course a little inconvenient, the module is so minimal that this should suffice.

For instructions read comment (strictly a string) in file!